requirements.txt
customtkinter==5.1.3
Pillow==9.5.0
praw==7.7.0
pyglet==2.0.7
pytz==2023.3
